requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.11
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv==1.0.0
pandas>=2.2.0
groq==0.4.1
python-dateutil==2.8.2

